Added Test button functions
* Test buttons work, you can test only those boards which are connected
to the teensy.
* Added API function, improved RPC docs [Tested]
* Fixes "port name change"
  - Say, teensy was connected on /dev/ttyACM0 and then the GUI flashed
it, the teensy would reconnect at /dev/ttyACM1 or whatever -- not ACM0.
That's because the GUI does not close the ACM0 port before flashing the
device.
  - This commit fixes that by giving the Teensy a (runtime-configurable)
amount of timeout (2 sec default) to reconnect on the old port.
  - THIS FIX IS NOT FOOL-PROOF AND WILL WORK WELL IN CASES WHERE Teensy
IS THE ONLY DEVICE BEING PROGRAMMED ON THE SERIAL PORT.
  - because we can't always gurantee the teensy would reconnect to the same
port it was previously on (but this solution assumes that).
  - Best Fix will require auto-detection of teensy on serial ports (not
supported by open-frameworks serial lib).